+++ src/elBSP.h
@@ -56,7 +56,7 @@ public:
 	class TempNode;
 private:
 	TempNode*		m_hierarchy;
-	unsigned int*	m_list;
+	uintptr_t*      m_list;
 	AABB			m_aabb;
 };
